requirements.txt
numpy
pandas
numba
matplotlib
pyyaml